refactor(cli): split sync workflow phases
Address review findings by moving command help and option metadata out of the
Typer adapters and splitting pull, lock, verification, and partial phases.

Use phase and file context objects so every helper has four or fewer arguments.
Keep each module below the CodeScene size and mean-complexity thresholds.
