+++ src/envdrift/cli.py
@@ -19,7 +19,7 @@ from envdrift.cli_commands.init_cmd import init as init_cmd
 from envdrift.cli_commands.install import install_app
 from envdrift.cli_commands.partial import pull_cmd as pull_partial_cmd
 from envdrift.cli_commands.partial import push as push_cmd
-from envdrift.cli_commands.sync import lock, pull, sync
+from envdrift.cli_commands.sync import LOCK_HELP, PULL_HELP, lock, pull, sync
 from envdrift.cli_commands.validate import validate
 from envdrift.cli_commands.vault import VAULT_PULL_HELP, VAULT_PUSH_HELP, vault_pull, vault_push
 from envdrift.cli_commands.version import version as version_cmd
@@ -89,8 +89,8 @@ app.command("init")(init_cmd)
 app.command()(guard)
 app.command()(hook)
 app.command()(sync)
-app.command()(pull)
-app.command()(lock)
+app.command(help=PULL_HELP)(pull)
+app.command(help=LOCK_HELP)(lock)
 app.command("vault-push", help=VAULT_PUSH_HELP)(vault_push)
 app.command("vault-pull", help=VAULT_PULL_HELP)(vault_pull)
 app.command()(version_cmd)
